# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "平行校园 Parallel Campus",
    "description": "知乎黑客松 2026 · 校园新锐季。公开数据 → 可编辑人格 → 分身在虚拟校园里替你先活一天。",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/persona/generate": {
      "post": {
        "tags": [
          "persona"
        ],
        "summary": "Generate",
        "description": "同步生成人格（服务端 60s 超时）。已有且非 force → 直接返回。",
        "operationId": "generate_api_persona_generate_post",
        "parameters": [
          {
            "name": "force",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Force"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonaResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/persona": {
      "get": {
        "tags": [
          "persona"
        ],
        "summary": "Get Persona",
        "operationId": "get_persona_api_persona_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonaResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "persona"
        ],
        "summary": "Put Persona",
        "operationId": "put_persona_api_persona_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PersonaUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PersonaResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/persona/deploy": {
      "post": {
        "tags": [
          "persona"
        ],
        "summary": "Deploy",
        "description": "投放到校园（US-05）。已投放且非 redeploy → 409。",
        "operationId": "deploy_api_persona_deploy_post",
        "parameters": [
          {
            "name": "redeploy",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Redeploy"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeployRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Deploy Api Persona Deploy Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CampusIdentity": {
        "properties": {
          "major": {
            "type": "string",
            "maxLength": 20,
            "title": "Major",
            "default": "未定"
          },
          "grade": {
            "type": "string",
            "enum": [
              "大一",
              "大二",
              "大三",
              "大四",
              "研一",
              "研二",
              "研三"
            ],
            "title": "Grade",
            "default": "大二"
          },
          "club": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 20
              },
              {
                "type": "null"
              }
            ],
            "title": "Club"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "CampusIdentity"
      },
...
      "DeployRequest": {
        "properties": {
          "display_name": {
            "type": "string",
            "maxLength": 24,
            "minLength": 1,
            "title": "Display Name"
          },
          "avatar_key": {
            "type": "string",
            "title": "Avatar Key",
            "default": "av_01"
          },
          "appearance": {
            "type": "string",
            "maxLength": 80,
            "title": "Appearance",
            "default": ""
          },
          "major": {
            "type": "string",
            "maxLength": 20,
            "title": "Major",
            "default": ""
          },
          "grade": {
            "type": "string",
            "enum": [
              "大一",
              "大二",
              "大三",
              "大四",
              "研一",
              "研二",
              "研三"
            ],
            "title": "Grade",
            "default": "大二"
          },
          "club": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 20
              },
              {
                "type": "null"
              }
            ],
            "title": "Club"
          }
        },
        "type": "object",
        "required": [
          "display_name"
        ],
        "title": "DeployRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Interest": {
        "properties": {
          "topic": {
            "type": "string",
            "maxLength": 12,
            "title": "Topic"
          },
          "weight": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Weight"
          },
          "evidence": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Evidence"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "topic",
          "weight"
        ],
        "title": "Interest"
      },
...
      "PersonaFile": {
        "properties": {
          "display_name": {
            "type": "string",
            "maxLength": 24,
            "minLength": 1,
            "title": "Display Name"
          },
          "archetype": {
            "type": "string",
            "maxLength": 20,
            "title": "Archetype"
          },
          "mbti_like": {
            "additionalProperties": {
              "type": "number"
            },
            "propertyNames": {
              "enum": [
                "E_I",
                "S_N",
                "T_F",
                "J_P"
              ]
            },
            "type": "object",
            "title": "Mbti Like"
          },
          "big_five": {
            "additionalProperties": {
              "type": "number"
            },
            "propertyNames": {
              "enum": [
                "O",
                "C",
                "E",
                "A",
                "N"
              ]
            },
            "type": "object",
            "title": "Big Five"
          },
          "interests": {
            "items": {
              "$ref": "#/components/schemas/Interest"
            },
            "type": "array",
            "maxItems": 8,
            "minItems": 3,
            "title": "Interests"
          },
          "stances": {
            "items": {
              "$ref": "#/components/schemas/Stance"
            },
            "type": "array",
            "maxItems": 5,
            "title": "Stances"
          },
          "speaking_style": {
            "$ref": "#/components/schemas/SpeakingStyle"
          },
          "values": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Values"
          },
          "social": {
            "$ref": "#/components/schemas/Social"
          },
          "campus_identity": {
            "$ref": "#/components/schemas/CampusIdentity"
          },
          "appearance": {
            "type": "string",
            "maxLength": 80,
            "title": "Appearance",
            "default": ""
          },
          "summary": {
            "type": "string",
            "maxLength": 400,
            "minLength": 1,
            "title": "Summary"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "display_name",
          "archetype",
          "mbti_like",
          "big_five",
          "interests",
          "summary"
        ],
        "title": "PersonaFile"
      },
      "PersonaResponse": {
        "properties": {
          "file": {
            "$ref": "#/components/schemas/PersonaFile"
          },
          "thin": {
            "type": "boolean",
            "title": "Thin",
            "default": false
          },
          "version": {
            "type": "integer",
            "title": "Version",
            "default": 1
          },
          "source_stats": {
            "$ref": "#/components/schemas/SourceStats"
          },
          "confirmed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confirmed At"
          },
          "generated_left_today": {
            "type": "integer",
            "title": "Generated Left Today",
            "default": 3
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "PersonaResponse"
      },
      "PersonaUpdateRequest": {
        "properties": {
          "file": {
            "$ref": "#/components/schemas/PersonaFile"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "PersonaUpdateRequest"
      },
...
      "Social": {
        "properties": {
          "initiative": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Initiative",
            "default": 0.5
          },
          "group_pref": {
            "type": "string",
            "enum": [
              "独处",
              "小圈子",
              "广交"
            ],
            "title": "Group Pref",
            "default": "小圈子"
          },
          "avoid_topics": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Avoid Topics"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "Social"
      },
      "SourceStats": {
        "properties": {
          "contents": {
            "type": "integer",
            "title": "Contents",
            "default": 0
          },
          "followees": {
            "type": "integer",
            "title": "Followees",
            "default": 0
          },
          "favlists": {
            "type": "integer",
            "title": "Favlists",
            "default": 0
          },
          "saved_items": {
            "type": "integer",
            "title": "Saved Items",
            "default": 0
          },
          "failed": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Failed"
          }
        },
        "type": "object",
        "title": "SourceStats"
      },
...
      "SpeakingStyle": {
        "properties": {
          "tone": {
            "type": "string",
            "maxLength": 30,
            "title": "Tone",
            "default": "自然"
          },
          "emoji": {
            "type": "boolean",
            "title": "Emoji",
            "default": false
          },
          "length": {
            "type": "string",
            "enum": [
              "短",
              "中",
              "长"
            ],
            "title": "Length",
            "default": "中"
          },
          "catchphrases": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Catchphrases"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "SpeakingStyle"
      },
      "Stance": {
        "properties": {
          "text": {
            "type": "string",
            "maxLength": 40,
            "title": "Text"
          },
          "evidence": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Evidence"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "text"
        ],
        "title": "Stance"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}